refactor: widgets 폴더 밑에 있는 폴더들(ChannelInfo/Picker/Widget)을 ChannelSpace 폴더로 병합 (#8)
