+++ src/widgets/ChannelSpace/ChannelInfo/index.ts
