+++ src/widgets/ChannelSpace/ChannelInfo/ui/ChannelInfo.tsx
