+++ src/widgets/ChannelSpace/ChannelInfo/ui/styles.ts
