+++ src/widgets/ChannelSpace/ChannelPicker/ui/ChannelPicker.tsx
