+++ src/widgets/ChannelSpace/ChannelPicker/ui/PickerStyles.ts
