+++ src/widgets/ChannelSpace/ChannelWidget/index.ts
