+++ src/widgets/ChannelSpace/ChannelWidget/ui/ChannelStyles.ts
