+++ src/widgets/ChannelSpace/ChannelWidget/ui/MemberList.tsx
